refactor: 번개 카드 리스트 중복 코드 분리
