+++ src/components/BunageMain/MainBungaeList.js
@@ -1,7 +1,5 @@
 import styled from "styled-components";
 
-import * as bungaeCardUtil from "../../@utils/bungaeCard";
-import BungaeCard from "../BungaeCard";
 import BungaeListContent from "../PageContent/BungaeListContent";
 import SimplePagination from "../UI/SimplePagination";
 
@@ -32,40 +30,7 @@ function NewBungaeList({ enHeading, koHeading, bungaeList }) {
         </div>
         <SimplePagination />
       </StyledUpperContent>
-      <BungaeListContent>
-        {bungaeList.map(
-          ({
-            id,
-            owner,
-            title,
-            location,
-            createdAt,
-            meetingAt,
-            numberOfParticipants,
-            numberOfRecruits
-          }) => {
-            const status = bungaeCardUtil.getBungaeStatus(createdAt, meetingAt);
-            const place = location.city + " " + location.state;
-            const time = bungaeCardUtil.getMeetingTime(meetingAt);
-            const duration = bungaeCardUtil.getBungaeDuration(meetingAt);
-            return (
-              <BungaeCard
-                key={id}
-                id={id}
-                status={status}
-                place={place}
-                time={time}
-                title={title}
-                emoji={owner.emoji}
-                nickname={owner.nickname}
-                numberOfParticipants={numberOfParticipants}
-                numberOfRecruits={numberOfRecruits}
-                duration={duration}
-              />
-            );
-          }
-        )}
-      </BungaeListContent>
+      <BungaeListContent bungaeList={bungaeList} />
     </>
   );
 }
