+++ src/components/Profile/UserBungaeList.js
@@ -1,7 +1,5 @@
 import styled from "styled-components";
 
-import * as bungaeCardUtil from "../../@utils/bungaeCard";
-import BungaeCard from "../BungaeCard";
 import BungaeListContent from "../PageContent/BungaeListContent";
 import SortTab from "../UI/SortTab";
 
@@ -19,40 +17,7 @@ function UserBungaeList({ sortPathname, switchTab, tabMenu, bungaeList }) {
         switchTab={switchTab}
         tabMenu={tabMenu}
       />
-      <BungaeListContent>
-        {bungaeList.map(
-          ({
-            id,
-            owner,
-            title,
-            location,
-            createdAt,
-            meetingAt,
-            numberOfParticipants,
-            numberOfRecruits
-          }) => {
-            const status = bungaeCardUtil.getBungaeStatus(createdAt, meetingAt);
-            const place = location.city + " " + location.state;
-            const time = bungaeCardUtil.getMeetingTime(meetingAt);
-            const duration = bungaeCardUtil.getBungaeDuration(meetingAt);
-            return (
-              <BungaeCard
-                key={id}
-                id={id}
-                status={status}
-                place={place}
-                time={time}
-                title={title}
-                emoji={owner.emoji}
-                nickname={owner.nickname}
-                numberOfParticipants={numberOfParticipants}
-                numberOfRecruits={numberOfRecruits}
-                duration={duration}
-              />
-            );
-          }
-        )}
-      </BungaeListContent>
+      <BungaeListContent bungaeList={bungaeList} />
     </StyledUserBungaeList>
   );
 }
